+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/qutip/QuantumToolbox.jl/tree/main)
 
+- Fix `cite()` bibtex output. ([#552])
+
 ## [v0.36.0]
 Release date: 2025-09-29
 
@@ -323,3 +325,4 @@ Release date: 2024-11-13
 [#539]: https://github.com/qutip/QuantumToolbox.jl/issues/539
 [#544]: https://github.com/qutip/QuantumToolbox.jl/issues/544
 [#546]: https://github.com/qutip/QuantumToolbox.jl/issues/546
+[#552]: https://github.com/qutip/QuantumToolbox.jl/issues/552