+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix `cite()` bibtex output. ([#552])
 - Add `qeye_like` and `qzero_like`, which are synonyms of `one` and `zero`. ([#555])
+- Add steadystate and DSF benchmarks. The `SteadyStateODESOlver` tolerances are lowered to `terminate_reltol=1e-4` and `terminate_abstol=1e-6` to improve speed at the cost of accuracy. ([#557])
 
 ## [v0.36.0]
 Release date: 2025-09-29
@@ -328,3 +329,4 @@ Release date: 2024-11-13
 [#546]: https://github.com/qutip/QuantumToolbox.jl/issues/546
 [#552]: https://github.com/qutip/QuantumToolbox.jl/issues/552
 [#555]: https://github.com/qutip/QuantumToolbox.jl/issues/555
+[#557]: https://github.com/qutip/QuantumToolbox.jl/issues/557